+++ crates/cubecl-core/src/prelude.rs
@@ -5,22 +5,23 @@ pub use crate::compute::{
     CompiledKernel, CubeKernel, KernelBuilder, KernelDefinition, KernelLauncher, KernelTask,
 };
 pub use crate::frontend::cmma;
-pub use crate::frontend::{branch::*, synchronization::*};
-pub use crate::runtime::Runtime;
-
 /// Elements
 pub use crate::frontend::{
     Array, ArrayHandleRef, Atomic, Float, FloatExpand, LaunchArg, NumericExpand, Slice, SliceMut,
     Tensor, TensorArg,
 };
+pub use crate::frontend::{branch::*, synchronization::*};
 pub use crate::pod::CubeElement;
+pub use crate::runtime::Runtime;
 
 /// Export plane operations.
 pub use crate::frontend::{plane_all, plane_max, plane_min, plane_prod, plane_sum};
 pub use cubecl_runtime::client::ComputeClient;
+pub use cubecl_runtime::id::KernelId;
+pub use cubecl_runtime::kernel::KernelMetadata;
 pub use cubecl_runtime::server::CubeCount;
 
 pub use crate::frontend::*;
 pub use crate::{comment, comptime, comptime_type, derive_cube_comptime, terminate};
-pub use cubecl_common::{CubeDim, ExecutionMode, KernelMetadata, flex32, tf32};
+pub use cubecl_common::{CubeDim, ExecutionMode, flex32, tf32};
 pub use cubecl_ir::Scope;
