+++ crates/cubecl-runtime/tests/dummy/kernel.rs
@@ -1,11 +1,10 @@
-use cubecl_common::id::KernelId;
-use cubecl_runtime::storage::BytesResource;
+use cubecl_runtime::{id::KernelId, storage::BytesResource};
 
 /// The DummyKernel trait should be implemented for every supported operation
 pub trait DummyKernel: Sync + Send + 'static + core::fmt::Debug {
     fn compute(&self, resources: &mut [&BytesResource]);
 
-    fn id(&self) -> cubecl_common::id::KernelId;
+    fn id(&self) -> KernelId;
 }
 
 /// Contains the algorithm for element-wise addition
@@ -27,7 +26,7 @@ impl DummyKernel for DummyElementwiseAddition {
         }
     }
 
-    fn id(&self) -> cubecl_common::id::KernelId {
+    fn id(&self) -> KernelId {
         KernelId::new::<Self>()
     }
 }
